moved paging files to a single folder
