+++ Makefile
@@ -1,4 +1,4 @@
-    OBJECTS = loader.o kmain.o common.o drivers/io.o drivers/framebuffer.o drivers/serial_port.o segmentation/gdt.o segmentation/memory_segments.o interrupts/idt.o interrupts/interrupt_handlers.o interrupts/interrupts.o interrupts/keyboard.o interrupts/pic.o user_mode/start_program.o paging_enable.o paging.o kheap.o
+    OBJECTS = loader.o kmain.o common.o drivers/io.o drivers/framebuffer.o drivers/serial_port.o segmentation/gdt.o segmentation/memory_segments.o interrupts/idt.o interrupts/interrupt_handlers.o interrupts/interrupts.o interrupts/keyboard.o interrupts/pic.o user_mode/start_program.o paging/paging_enable.o paging/paging.o paging/kheap.o
     CC = gcc
     CFLAGS = -m32 -nostdlib -nostdinc -fno-builtin -fno-stack-protector \
              -nostartfiles -nodefaultlibs -Wall -Wextra -Werror -c -masm=intel
