+++ paging/paging.c
@@ -1,5 +1,5 @@
 #include "paging.h"
-#include "drivers/serial_port.h" // pour afficher un msg
+#include "../drivers/serial_port.h" // pour afficher un msg
